+++ lib/belt/route_dsl.rb
@@ -75,8 +75,11 @@ module Belt
       resource_options = options.merge(route_type: :resources)
       actions = @gateway.send(:determine_actions, options)
 
-      add_nested_resource_routes(resource_name, param_name, resource_options, actions)
-      return unless block
+      # When there's a block (nested resources or custom member actions), use {param_name}
+      # for member routes to avoid API Gateway sibling path parameter conflicts.
+      has_nested = !block.nil?
+      add_nested_resource_routes(resource_name, param_name, resource_options, actions, use_param_name: has_nested)
+      return unless has_nested
 
       nested_member_prefix = "#{@prefix}/#{resource_name}/{#{param_name}}"
       nested_collection_prefix = "#{@prefix}/#{resource_name}"
@@ -251,7 +254,11 @@ module Belt
       end
     end
 
-    def add_nested_resource_routes(resource_name, param_name, resource_options, actions)
+    def add_nested_resource_routes(resource_name, param_name, resource_options, actions, use_param_name: false)
+      # Member routes (show/update/destroy) use {id} by default (Rails convention).
+      # When nested resources or custom member actions exist (use_param_name: true),
+      # use {param_name} instead to avoid API Gateway sibling path parameter conflicts.
+      member_param = use_param_name ? param_name : 'id'
       if actions.include?(:index)
         @gateway.send(:add_route, :get, "#{@prefix}/#{resource_name}",
                       resolve_request_model_for(resource_options, :index))
@@ -261,16 +268,16 @@ module Belt
                       resolve_request_model_for(resource_options, :create))
       end
       if actions.include?(:show)
-        @gateway.send(:add_route, :get, "#{@prefix}/#{resource_name}/{#{param_name}}",
+        @gateway.send(:add_route, :get, "#{@prefix}/#{resource_name}/{#{member_param}}",
                       resolve_request_model_for(resource_options, :show))
       end
       if actions.include?(:update)
-        @gateway.send(:add_route, :put, "#{@prefix}/#{resource_name}/{#{param_name}}",
+        @gateway.send(:add_route, :put, "#{@prefix}/#{resource_name}/{#{member_param}}",
                       resolve_request_model_for(resource_options, :update))
       end
       return unless actions.include?(:destroy)
 
-      @gateway.send(:add_route, :delete, "#{@prefix}/#{resource_name}/{#{param_name}}",
+      @gateway.send(:add_route, :delete, "#{@prefix}/#{resource_name}/{#{member_param}}",
                     resolve_request_model_for(resource_options, :destroy))
     end
 
@@ -415,9 +422,12 @@ module Belt
       resource_options = options.merge(route_type: :resources)
       actions = determine_actions(options)
 
-      add_resource_routes(resource_name, param_name, resource_options, actions)
+      # When there are nested resources, use {param_name} for member routes to avoid
+      # API Gateway sibling path parameter conflicts. Without nested resources, use {id}.
+      has_nested = block_given?
+      add_resource_routes(resource_name, param_name, resource_options, actions, use_param_name: has_nested)
 
-      return unless block_given?
+      return unless has_nested
 
       collection_prefix = "/#{resource_name}"
       member_prefix = "/#{resource_name}/{#{param_name}}"
@@ -496,7 +506,11 @@ module Belt
       options.merge(tables: [resource_name.to_sym])
     end
 
-    def add_resource_routes(resource_name, param_name, resource_options, actions)
+    def add_resource_routes(resource_name, param_name, resource_options, actions, use_param_name: false)
+      # Member routes (show/update/destroy) use {id} by default (Rails convention).
+      # When nested resources exist (use_param_name: true), use {param_name} instead
+      # to avoid API Gateway sibling path parameter conflicts.
+      member_param = use_param_name ? param_name : 'id'
       if actions.include?(:index)
         add_route(:get, "/#{resource_name}",
                   resolve_request_model_for(resource_options, :index))
@@ -506,16 +520,16 @@ module Belt
                   resolve_request_model_for(resource_options, :create))
       end
       if actions.include?(:show)
-        add_route(:get, "/#{resource_name}/{#{param_name}}",
+        add_route(:get, "/#{resource_name}/{#{member_param}}",
                   resolve_request_model_for(resource_options, :show))
       end
       if actions.include?(:update)
-        add_route(:put, "/#{resource_name}/{#{param_name}}",
+        add_route(:put, "/#{resource_name}/{#{member_param}}",
                   resolve_request_model_for(resource_options, :update))
       end
       return unless actions.include?(:destroy)
 
-      add_route(:delete, "/#{resource_name}/{#{param_name}}", resolve_request_model_for(resource_options, :destroy))
+      add_route(:delete, "/#{resource_name}/{#{member_param}}", resolve_request_model_for(resource_options, :destroy))
     end
 
     def resolve_request_model_for(options, action)
@@ -797,7 +811,11 @@ module Belt
         end
       end
 
-      def add_scoped_resource_routes(resource_name, param_name, resource_options, actions)
+      def add_scoped_resource_routes(resource_name, param_name, resource_options, actions, use_param_name: false)
+        # Member routes (show/update/destroy) use {id} by default (Rails convention).
+        # When nested resources exist (use_param_name: true), use {param_name} instead
+        # to avoid API Gateway sibling path parameter conflicts.
+        member_param = use_param_name ? param_name : 'id'
         if actions.include?(:index)
           @gateway.send(:add_route, :get, build_path("/#{resource_name}"),
                         resolve_request_model_for(resource_options, :index))
@@ -807,16 +825,16 @@ module Belt
                         resolve_request_model_for(resource_options, :create))
         end
         if actions.include?(:show)
-          @gateway.send(:add_route, :get, build_path("/#{resource_name}/{#{param_name}}"),
+          @gateway.send(:add_route, :get, build_path("/#{resource_name}/{#{member_param}}"),
                         resolve_request_model_for(resource_options, :show))
         end
         if actions.include?(:update)
-          @gateway.send(:add_route, :put, build_path("/#{resource_name}/{#{param_name}}"),
+          @gateway.send(:add_route, :put, build_path("/#{resource_name}/{#{member_param}}"),
                         resolve_request_model_for(resource_options, :update))
         end
         return unless actions.include?(:destroy)
 
-        @gateway.send(:add_route, :delete, build_path("/#{resource_name}/{#{param_name}}"),
+        @gateway.send(:add_route, :delete, build_path("/#{resource_name}/{#{member_param}}"),
                       resolve_request_model_for(resource_options, :destroy))
       end
 
@@ -839,7 +857,10 @@ module Belt
         resource_options = options.merge(route_type: :resources, controller: controller)
         actions = determine_scoped_actions(options)
 
-        add_scoped_resource_routes(resource_name, param_name, resource_options, actions)
+        # When there are nested resources, use {param_name} for member routes to avoid
+        # API Gateway sibling path parameter conflicts. Without nested resources, use {id}.
+        has_nested = !block.nil?
+        add_scoped_resource_routes(resource_name, param_name, resource_options, actions, use_param_name: has_nested)
         build_nested_resource_block(resource_name, param_name, options, controller, &block) if block
       end